graphiq/flask-api/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.0
asyncio==3.4.3
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==6.2.13
faiss-cpu==1.8.0.post1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-anthropic==0.2.3
langchain-community==0.3.1
langchain-core==0.3.13
langchain-huggingface==0.1.0
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langsmith==0.1.129
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.22.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
primp==0.6.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.9.0
pypdf==5.0.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
twilio==9.3.2
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.13.1